--- a/src/SearchableMap/fuzzySearch.ts
+++ b/src/SearchableMap/fuzzySearch.ts
@@ -21,7 +21,7 @@ export const fuzzySearch = <T = any>(node: RadixTree<T>, query: string, maxDista
   const maxLength = n + maxDistance
 
   // Fill first matrix row with consecutive numbers 0 1 2 3 ... (n - 1)
-  const matrix = new Uint8Array(maxLength * maxLength)
+  const matrix = new Uint8Array(maxLength * n)
   for (let i = 0; i < n; i++) matrix[i] = i
 
   recurse(
@@ -38,21 +38,53 @@ export const fuzzySearch = <T = any>(node: RadixTree<T>, query: string, maxDista
   return results
 }
 
-const recurse = <T = any>(node: RadixTree<T>, query: string, maxDistance: number, results: FuzzyResults<T>, matrix: Uint8Array, offset: number, n: number, prefix: string): void => {
+// Modified version of http://stevehanov.ca/blog/?id=114
+
+// This builds a Levenshtein matrix for a given query and continuously updates
+// it for nodes in the radix tree that fall within the given maximum edit
+// distance. Keeping the same matrix around is beneficial especially for larger
+// edit distances.
+//
+//           k   a   t   e   <-- query
+//       0   1   2   3   4
+//   c   1   1   2   3   4
+//   a   2   2   1   2   3
+//   t   3   3   2   1  [2]  <-- edit distance
+//   ^
+//   ^ term in radix tree, rows are added and removed as needed
+
+const recurse = <T = any>(
+  node: RadixTree<T>,
+  query: string,
+  maxDistance: number,
+  results: FuzzyResults<T>,
+  matrix: Uint8Array,
+  offset: number,
+  n: number,
+  prefix: string
+): void => {
   key: for (const key of node.keys()) {
     if (key === LEAF) {
+      // We've reached a leaf node. Check if the edit distance acceptable and
+      // store the result if it is.
       const distance = matrix[offset - 1]
       if (distance <= maxDistance) {
         results[prefix] = [node.get(key) as T, distance]
       }
     } else {
+      // Iterate over all characters in the key. Update the Levenshtein matrix
+      // and check if the minimum distance in the last row is still within the
+      // maximum edit distance. If it is, we can recurse over all child nodes.
       for (let i = 0; i < key.length; i++) {
         const char = key[i]
         const thisRowOffset = offset + n * i
         const prevRowOffset = thisRowOffset - n
 
+        // Set the first column based on the previous row, and initialize the
+        // minimum distance in the current row.
         let minDistance = matrix[thisRowOffset] = matrix[prevRowOffset] + 1
 
+        // Iterate over remaining columns (characters in the query).
         for (let j = 0; j < n - 1; j++) {
           const different = char !== query[j]
 
@@ -68,7 +100,8 @@ const recurse = <T = any>(node: RadixTree<T>, query: string, maxDistance: number
           if (dist < minDistance) minDistance = dist
         }
 
-        // Because distance will never decrease, we can stop here.
+        // Because distance will never decrease, we can stop. There will be no
+        // matching child nodes.
         if (minDistance > maxDistance) {
           continue key
         }