@lucaong 2022-02-17
You are on fire @rolftimmermans ! 🔥 

Yes, fuzzy match bound by edit distance on a trie is complex. Other data structures (like trigram indexes) would be theoretically faster, but MiniSearch has to optimize for memory utilization. Luckily, in practice, the maximum edit distance is bound to a small number, and when searching on the client side it makes sense to trade computation for a smaller space utilization, so a single general-purpose index data structure is preferred. I elaborate on that a bit in [this blog post](https://lucaongaro.eu/blog/2019/01/30/minisearch-client-side-fulltext-search-engine.html).

It's great to see such significant gains! I was hoping indeed that using this dynamic programming technique would get MiniSearch closer to the theoretical optimum.

Thanks again!!