@lucaong 2022-02-17
@rolftimmermans this and all your outstanding contributions are now released as part of MiniSearch `v4.0.0`. I took the liberty of mentioning you in the CHANGELOG. Thanks again!

I will now upgrade my production apps to use the new version, looking forward to see the improvements :)